Move constants to constants.go file
